# db: fix NewIter regression resulting in extra memtable levels